Adding logic for create an auction with vehicle

+++ src/main/java/com/cinema/starwars/controllers/VehicleController.java
@@ -22,7 +22,7 @@ public class VehicleController {
     }
 
     @GetMapping("one/{id}")
-    public Vehicle getOneVehicleFromGarage(@PathVariable int id){
+    public Vehicle getOneVehicleFromGarage(@PathVariable Long id){
 
         return vehicleService.getOneVehicleFromGarage(id);
     }
